Converted "RGB Distance" to "Cube Distance" between 0 and 100
The raw RGB distance depends on the number of bits per color. This
can be normalized to a maximum of (1,1,1). However this does not
compare well with the CIE deltas which range from 0 to 100. So the
former "RGB Distance" was scaled to produce a value from 0 to 100.
This is currently called "Cube Distance".